helpful-finetuning/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helpful-finetuning"
version = "0.1.0"
description = "Stage 2: Helpful Response Fine-tuning (QLoRA on Gemma-7B-IT) with evaluation and Streamlit comparison"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.40.2,<5.0.0",
    "datasets>=2.19.1,<3.0.0",
    "peft>=0.7.0",
    "accelerate>=0.28.0",
    "wandb>=0.17.0,<1.0.0",
    "streamlit>=1.33.0,<2.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "numpy<2.0.0",
    "evaluate>=0.4.2,<0.6.0",
    "scikit-learn>=1.3.0,<2.0.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.4.0,<25.0.0",
    "ruff>=0.4.0,<0.5.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
